+++ examples/react-spa/backend/backend/settings.py
@@ -137,6 +137,7 @@ HEADLESS_FRONTEND_URLS = {
     "account_reset_password": "/account/password/reset",
     "account_reset_password_from_key": "/account/password/reset/key/{key}",
     "account_signup": "/account/signup",
+    "socialaccount_login_error": "/account/provider/callback",
 }
 
 try: